mcp_hexagon/pyproject.toml
[project]
name = "mcp-hexagon"
version = "0.1.0"
description = "MCP for Hexagon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.10.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[project.scripts]
hexagon-mcp = "hexagon_mcp_server.server:main"

[tool.uv]
package = true